Progetto IOT/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
wtforms
google-cloud-pubsub
validators
python-Levenshtein
numpy
pandas
matplotlib
seaborn
sklearn
flask_sqlalchemy
googlemaps